Start working on custom ast based remask-shiki-twoslash plugin
