+++ apps/docs/docs/react-imperative-hook/index.mdx
@@ -30,6 +30,12 @@ i.e. if you want separate rendering for modals and toasts.
 
 See the [Primitives](#primitives) section for more information on each primitive.
 
+```tsx twoslash
+console.log("testing");
+const foo = 123;
+//    ^?
+```
+
 ```tsx
 // imperative.ts
