+++ apps/docs/package.json
@@ -25,12 +25,10 @@
     "@mdx-js/react": "^2.3.0",
     "@types/node": "^20.5.8",
     "@types/react": "18.2.21",
-    "@yas/tsconfig": "workspace:*",
     "@yas/env": "workspace:*",
+    "@yas/tsconfig": "workspace:*",
     "docusaurus-plugin-sass": "^0.2.5",
-    "prism-react-renderer": "2.0.6",
-    "rehype-raw": "^7.0.0",
-    "remark-shiki-twoslash": "^3.1.3",
+    "remark-shiki-twoslash": "workspace:*",
     "sass": "^1.66.1",
     "sass-loader": "^13.3.2",
     "shiki-twoslash-loader": "workspace:*",
