+++ packages/remark-shiki-twoslash/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "remark-shiki-twoslash",
+  "version": "0.0.1",
+  "main": "./index.js",
+  "devDependencies": {
+    "shiki-twoslash-loader": "workspace:*"
+  }
+}
